Add script for accessing files in Google Drive
Add script for accessing files in Google Drive so we can use Drive to store and manage our raw data. This will enusure that each group member use the same raw data and limit chance of those errors. Also this will provide a common source api for the data so that the scripts we develop work in every enviroment.